fix(#4158): make render_field_marshal Cassandra's ignoreFreezing, not a hand-rolled rule
The registry-aware sibling renderer in the SAME emit path carried its own
approximation of the freeze rule -- 'drop the FrozenType wrapper only for a
DIRECT UDT field, keep it for a frozen collection' (#1020 roborev Finding 2).
Cassandra's actual rule is one flag: every composite stringifies its subtypes
with 'ignoreFreezing || !isMultiCell' (UserType.java:444, MapType.java:317,
SetType.java:193, ListType.java:203) and TupleType always passes true
(TupleType.java:559). So a frozen UserType suppresses ALL its children's
wrappers, not just a UDT child.

Measured against the unfixed renderer, a frozen UDT's fields came out as
  fl:  FrozenType(ListType(Int32Type))        -- Cassandra: ListType(Int32Type)
  fs:  FrozenType(SetType(UTF8Type))          -- Cassandra: SetType(UTF8Type)
  tup: TupleType(Int32Type,FrozenType(ListType(Int32Type)))
                                              -- Cassandra: TupleType(Int32Type,ListType(Int32Type))
while the corpus-attested shapes (a nested UDT field bare inside
FrozenType(UserType(..)); ListType(FrozenType(UserType(..))) for a MULTICELL
list of frozen UDT; FrozenType(ListType(UserType(..))) for the frozen column)
are unchanged. It also removes the disagreement the string converter's #4158
fix would otherwise have opened between the two renderers.

render_field_marshal now takes ignore_freezing and shares the wrapper-eligibility
predicate with the string converter (stats_writer::marshal::takes_frozen_type_wrapper),
so there is ONE rule. is_direct_udt is gone with its special case. The tests now
say WHICH context they assert -- as_column (toString(false)) vs
as_frozen_udt_field -- because 'bare UserType' is right for a field and WRONG for
a column, and the old suite conflated them.

schema_helpers.rs was already 908 lines (over the 800 target), so its inline
tests move to schema_helpers/schema_helpers_tests.rs per the campsite rule; the
source is now 775 and the file-size ratchet stays PASS rather than OPT-OUT.
